+++ CHANGELOG.markdown
@@ -1,7 +1,7 @@
 ### 0.2.2 / 2012/02/13
 * :underscore option added
 * `Insensitive::KeyClashError` added for safer operations
-* Bug fix: Preserve default/default_proc on merge/replace
+* Bug fix: Preserve default/default_proc on merge/replace/insensitive
 * Subtle behavioral change in Insensitive#replace when called with an ordinary Hash
 
 ### 0.2.1 / 2012/02/10